More KF5 porting
Rename kdialogd to match the KDE Frameworks version.  Also version the
shared libraries so that they can be co-installed with the KDE4 version.

Look at KDE_SESSION_VERSION in kgtk-wrapper so that it can work
for both KDE4 and KF5.

Fix kgtk-warpper where it is called with an absolute path for the target
executable.